pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "chromadb>=1.5.5",
    "evaluate>=0.4.6",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "nltk>=3.9.3",
    "peft>=0.18.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "pypdf>=6.8.0",
    "pyyaml>=6.0.3",
    "ragas>=0.4.3",
    "rouge-score>=0.1.2",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]
